stable-diffusion-project/genLabs_frontend/requirements.txt
appnope==0.1.3
backcall==0.2.0
Bootstrap-Flask==2.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
firebase-admin==6.0.0
Flask==2.2.2
gcloud==0.18.3
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-secret-manager==2.12.6
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
jws==0.1.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.21.6
oauth2client==4.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
Pygments==2.13.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
pyzmq==24.0.1
requests==2.28.1
requests-toolbelt==0.10.0
rsa==4.9
six==1.16.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
